+++ packages/types/src/data.ts
@@ -265,3 +265,9 @@ export type SessionUrls = {
   readonly batch?: string
   readonly session?: string
 }
+
+export type LazyLoadOptions = {
+  scrollLength?: number
+  waitingTime?: number
+  pageHeight?: number
+}